sstables: move sstable_writer to separate header
This class is used in only few places and does not have to be included
everywhere sstable class is needed.
