+++ sstables/sstables.cc
@@ -43,6 +43,7 @@
 #include "writer.hh"
 #include "m_format_read_helpers.hh"
 #include "sstables.hh"
+#include "sstable_writer.hh"
 #include "metadata_collector.hh"
 #include "progress_monitor.hh"
 #include "compress.hh"
