+++ test/boost/sstable_datafile_test.cc
@@ -36,6 +36,7 @@
 #include "database.hh"
 #include "sstables/leveled_manifest.hh"
 #include "sstables/metadata_collector.hh"
+#include "sstables/sstable_writer.hh"
 #include <memory>
 #include "test/boost/sstable_test.hh"
 #include <seastar/core/seastar.hh>
